Restructure app files 📁
