+++ components/AppCard.tsx
