+++ tailwind.config.ts
@@ -6,7 +6,8 @@ const config: Config = {
 		hoverOnlyWhenSupported: true,
 	},
 	content: [
-		'./app/**/*.{js,ts,jsx,tsx,mdx}',
+		"./app/**/*.{js,ts,jsx,tsx,mdx}",
+		"./components/**/*.{js,ts,jsx,tsx,mdx}",
 		"!./node_modules", 
 	],
 	theme: {
