+++ tsconfig.json
@@ -20,15 +20,14 @@
     ],
     "paths": {
       "@/*": ["./*"],
-			"@styles/*": ["./app/_styles/*"],
-			"@public/*": ["./app/_public/*"],
-			"@components/*": ["./app/_components/*"],
+			"@styles/*": ["./styles/*"],
+			"@public/*": ["./public/*"],
+			"@components/*": ["./components/*"],
 			"@lib/*": ["./lib/*"],
 			"@dashboard/*": ["./app/(dashboard)/*"],
 			"@data/*": ["./app/_data/*"],
-			"@helpers/*": ["./app/_helpers/*"],
     },
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
-  "exclude": ["node_modules"],
+  "exclude": ["node_modules","prisma"],
 }
